refactor: refactor prioritypathogenview to prioritypathogensview (#607)
